Moved kernel_providers to its own file
Also moved related tests to this new file
